refactor(poly check, poly libs): extract poetry internals to separate module
